Update the admin role using the backend
