+++ src/components/Admin/Layout/AppBar.jsx
@@ -1,0 +1,85 @@
+import React from 'react';
+import {
+  AppBar,
+  CssBaseline,
+  IconButton,
+  Toolbar,
+  Typography,
+  Button,
+} from '@material-ui/core';
+import MenuIcon from '@material-ui/icons/Menu';
+import DropDown from '../../Layout/DropDownMenu';
+import { makeStyles } from '@material-ui/core/styles';
+import { Link } from 'react-router-dom';
+import { connect } from 'react-redux';
+
+const useStyles = makeStyles((theme) => {
+  return {
+    appBar: {
+      zIndex: theme.zIndex.drawer + 1,
+      background: theme.palette.primary[800],
+    },
+    appBarTitle: {
+      flexGrow: 1,
+    },
+    menuButton: {
+      marginRight: theme.spacing(2),
+      [theme.breakpoints.up('sm')]: {
+        display: 'none',
+      },
+    },
+    adminButton: {
+      marginRight: theme.spacing(3),
+    },
+  };
+});
+
+function MyAppBar(props) {
+  const classes = useStyles();
+  const [mobileOpen, setMobileOpen] = React.useState(false);
+
+  const handleDrawerToggle = () => {
+    setMobileOpen(!mobileOpen);
+  };
+
+  return (
+    <div>
+      <CssBaseline />
+      <AppBar position='fixed' className={classes.appBar}>
+        <Toolbar>
+          <IconButton
+            color='inherit'
+            aria-label='open drawer'
+            edge='start'
+            onClick={handleDrawerToggle}
+            className={classes.menuButton}
+          >
+            <MenuIcon />
+          </IconButton>
+          <Typography variant='h6' className={classes.appBarTitle} noWrap>
+            SCHEDU ADMIN
+          </Typography>
+          <Button
+            className={classes.adminButton}
+            component={Link}
+            to='/'
+            color='secondary'
+            variant='contained'
+          >
+            Back To User Panel
+          </Button>
+          <DropDown />
+        </Toolbar>
+      </AppBar>
+    </div>
+  );
+}
+
+const mapStateToProps = (state) => {
+  return {
+    auth: state.firebase.auth,
+    admin: state.admin,
+  };
+};
+
+export default connect(mapStateToProps)(MyAppBar);
