+++ src/store/actions/action-names.js
@@ -4,12 +4,12 @@ const auth = {
   SIGNOUT_SUCCESS: 'SIGNOUT_SUCCESS',
   REGISTER_SUCCESS: 'REGISTER_SUCCESS',
   REGISTER_ERROR: 'REGISTER_ERROR',
+  SET_USER_ADMIN_STATUS: 'SET_USER_ADMIN_STATUS',
 };
 
 const admin = {
   SEARCH_USER_BY_EMAIL_SUCCESS: 'SEARCH_USER_BY_EMAIL_SUCCESS',
   SEARCH_USER_BY_EMAIL_FAIL: 'SEARCH_USER_BY_EMAIL_FAIL',
-  SET_USER_ADMIN_STATUS: 'SET_USER_ADMIN_STATUS',
   UPDATE_USER_ADMIN_ROLE_SUCCESS: 'UPDATE_USER_ADMIN_ROLE_SUCCESS',
   UPDATE_USER_ADMIN_ROLE_FAIL: 'UPDATE_USER_ADMIN_ROLE_FAIL',
   ENABLE_DISABLE_USER_SUCCESS: 'ENABLE_DISABLE_USER_SUCCESS',
