+++ src/store/reducers/admin-reducer.js
@@ -3,7 +3,6 @@ import actionNames from './../actions/action-names';
 const {
   SEARCH_USER_BY_EMAIL_SUCCESS,
   SEARCH_USER_BY_EMAIL_FAIL,
-  SET_USER_ADMIN_STATUS,
   ENABLE_DISABLE_USER_SUCCESS,
   ENABLE_DISABLE_USER_FAIL,
   UPDATE_USER_ADMIN_ROLE_SUCCESS,
@@ -14,7 +13,6 @@ const {
 } = actionNames.admin;
 
 const initState = {
-  isAdmin: false,
   searchedUser: null,
   admins: [],
   adminError: null,
@@ -32,11 +30,6 @@ const adminReducer = (state = initState, action) => {
       return {
         ...state,
         adminError: action.data,
-      };
-    case SET_USER_ADMIN_STATUS:
-      return {
-        ...state,
-        isAdmin: action.isAdmin,
       };
     case ENABLE_DISABLE_USER_SUCCESS:
       return {
@@ -56,7 +49,7 @@ const adminReducer = (state = initState, action) => {
       return {
         ...state,
         notificationMessage: action.data.message,
-        searchedUser: action.data.user,
+        searchedUser: action.data,
       };
     case UPDATE_USER_ADMIN_ROLE_FAIL:
       return {
