+++ src/store/reducers/auth-reducer.js
@@ -6,10 +6,12 @@ const {
   SIGNOUT_SUCCESS,
   REGISTER_SUCCESS,
   REGISTER_ERROR,
+  SET_USER_ADMIN_STATUS,
 } = actionNames.auth;
 
 const initState = {
   authError: null,
+  isAdmin: false,
 };
 
 const authReducer = (state = initState, action) => {
@@ -34,8 +36,13 @@ const authReducer = (state = initState, action) => {
         ...state,
         authError: null,
       };
+    case SET_USER_ADMIN_STATUS:
+      return {
+        ...state,
+        isAdmin: action.isAdmin,
+      };
     case SIGNOUT_SUCCESS:
-      return state;
+      return { ...state, isAdmin: false };
     default:
       return state;
   }
